Update patch set 1
Patch Set 1: Code-Review+1

In TimerFactory.java, in addition to LacpWheelTimer only a timerWheel hashmap is present. Is 2 public classes required in the same file. For now, i am approving the change. Later either of this modification can be done. If the LacpWheelTimer is not required outside this file, the class can be made private. or the timerWheel hash map can be moved to the LacpMainModule.java.

Patch-set: 1
Reviewer: Gerrit User 3374 <3374@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Code-Review=+1