Update patch set 2
Patch Set 2:

The scalability of the LACP feature with the Lithium code base is claimed for 128 nodes with each node having 1 lag with 8 ports in it in the slow timeout mode.
The scalability was verified using mininet simulation switches.
Can you please verify if LACP supports the above given scalability numbers with the queue data structure change.

Patch-set: 2